move Component to its own package
